backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.0
# google-generativeai==0.3.2  # Removed - now using OpenRouter
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
jinja2==3.1.2
PyYAML==6.0.1 
playwright==1.40.0
pytest==7.4.3
pytest-html==4.1.1
selenium==4.15.2
webdriver-manager==4.0.1 